Add classes and data directories, replaced tkinter with command-line run.
This commit adds increased functionality to PlanB as now one can add specific
files, folders, and settings to the program via the .json files in the data
directory. These files and folders will be copied from their path to the path
specified in the settings.json file.

Although there is no support for tkinter, the program is more functional and
more simple. Once a user enters all of the files and folders they would like
to backup, the user can run the program as follows:

Within a terminal, Go to the projects root directory and type:

	python main.py

This will attempt to back up all of the user-specified data into the save
directory specified in the settings.json file.

Errors will occur if the program encounters read-only files! If this occurs,
I reccommend manually deleting the whole save directory and running the program again.