add updater.py and fix read-only file deletion.
updater.py houses the Updater class which handles all the messages that will be
printed to the user to inform them of the program's progress.

As of the last commit, read-only files could not be deleted by the program. This is
fixed by using a static method call that uses os.chmod to change the read-only flag.
The method then deletes the modified file as usual.